requirements.txt
asgiref==3.11.1
beautifulsoup4==4.14.3
Django==4.2.29
django-bootstrap-v5==1.0.11
django-environ==0.13.0
djangorestframework==3.17.1
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
